services/sim/requirements.txt
gekko>=1.0.6
nats-py>=2.4.0
protobuf>=4.21.0
python-dotenv>=1.0.0
numpy>=1.24.0
